Entrega n°3 código QR funcional
