+++ src/app/home/home.page.html
@@ -4,7 +4,7 @@
   </ion-toolbar>
 </ion-header>
 
-<ion-content [style.visibility]="estado_visible">
+<ion-content>
 
   
   <br>
@@ -25,9 +25,9 @@
       <hr>
       <ion-card-subtitle class="ion-text-center">Sistema de asistencia registrAPP</ion-card-subtitle>
       <hr>
-      <ion-item *ngIf="resultadoQR">
-        <ion-label class="ion-text-wrap">Resultado: {{ resultadoQR }}</ion-label>
-      </ion-item>
+      
+      
+      
     </ion-card>
   </div>
   <ion-img src='../../assets/icon/duoc.png' alt="DuocUC-ICO"></ion-img>
