+++ src/app/restore/restore.page.ts
@@ -87,38 +87,7 @@ export class RestorePage implements OnInit {
       
     }
   }
-  /*
-  restore(){//La función busca en el arreglo de usuarios si el usuario coincide y guarda los datos de este en la 
-            //constante user y la posición donde se encontraba en la variable index
-    const user = this.usuariosReg.find(user => user.username == this.username);
-    const index = this.usuariosReg.findIndex(user => user.username == this.username); 
-    
-    if (user){//Si el usuario existe se procedé a validar que las contraseñas ingresadas sean iguales
-      console.log(user)
-      if (this.password == this.password2) {//Si las contraseñas son iguales se reemplazan los datos del usuario antiguo del localStorage 
-                                            //manteniendo todos a excepción de la nueva contraseña ingresada
-        this.usuariosReg.splice(index, 1, {nombre: user?.nombre, apellido: user?.apellido, email: user?.email, tipo: user?.tipo, username: user?.username, password: this.password, carrera: user?.carrera, ramos: user?.ramos});
-        
-        localStorage.setItem('usuariosRegistrados', JSON.stringify(this.usuariosReg));
-        //Se actualiza el arreglo con todos los usuarios nuevamente en el localStorage
-        
-        this.mostrarAlerta3(true)
-        //Se muestra la confirmación del cambio exitoso de contraseña
-        
-      }else{//Si las contraseñas no son iguales se muestra la alerta correspondiente y se limpian los campos de estas
-        this.mostrarAlerta2(true)
-        
-        this.password= '';
-        this.password2= '';
-      }
-    }else{//Si el usuario no existe se invoca la alerta correspondiente y se limpian todos los campos
-      this.mostrarAlerta(true)
-      this.username= '';
-      this.password= '';
-      this.password2= '';
-    }
-  }
-  */
+  
   ngOnInit() {
     this.username= '';
     this.password= '';
